Added check to see if cache support is available. This requires https or overriding settings on the browser to allow the cache features. Also includes line spreading.
Caching speeds up file reloading especially after a refresh so that file does not need to be redownloaded. Also helps reduce the memory foot print since I no longer have to hold the file in memory for changing settings like render quality. I can reach back to the cache and pull the file in one shot. Since this requires https I believe it will only be supported on the Duet 3 + SBC for hosting.