Moved some ui elements around
I moved the controls into a slide out panel so the viewer would have more real estate. On top of that added some window resizing events to attempt to size the div to fit the available area better than it has and prevent the need for a scroll bar. the goal being that all the status information remains on screen.