Resolved an issue with gcode processing
There was an issue where commands like g90 and g91 were not properly processed because they do not have additional tokens associated with them. This was causing issue #2 to not go back to absolute mode after being set to relative. the issue only cropped up because a G91 in some sample gcode had additional gcode inline with the call.